requirements.txt
Django
django-allauth
psycopg2-binary
numpy
pandas
requests
PyJWT==2.3.0
cryptography